requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=0.24.2
matplotlib>=3.4.3
seaborn>=0.11.2
lime>=0.2.0.1
optuna>=2.10.0
plotly>=5.3.1
dash>=2.0.0
dash-bootstrap-components>=0.13.0
python-dotenv>=0.19.0
flask>=2.0.0
gunicorn>=20.1.0 